AI's weird, Chain Gun Kinda Works!
I got the chain gun to work, partially! Damage is being applied to the player, the AI broke a little and sometimes yeets away from the player

+++ Base Hack n' Slash Project/Assets/Scripts/HUDController.cs
@@ -9,13 +9,13 @@ public class HUDController : MonoBehaviour
     public Slider dashCooldown;
     public Slider hookCooldown;
 
-    public void SetHealthBarMaxValue(int value)
+    public void SetHealthBarMaxValue(float value)
     {
         healthBar.maxValue = value;
         healthBar.value = value;
     }
 
-    public void SetHealthBarValue(int value)
+    public void SetHealthBarValue(float value)
     {
         healthBar.value = value;
     }
